+++ wp_com_e2e/src/context.rs
@@ -41,6 +41,7 @@ impl TestContext {
             request_executor: Arc::new(ReqwestRequestExecutor::new(false, Duration::from_secs(60))),
             middleware_pipeline: Arc::new(WpApiMiddlewarePipeline::default()),
             app_notifier: Arc::new(EmptyAppNotifier),
+            language_provider: None,
         };
 
         let client = WpComApiClient::new(delegate.clone());